test(cypress): move config tests to integration suite
