+++ .github/workflows/instrumentation.yml
@@ -177,16 +177,6 @@ jobs:
       - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
       - uses: ./.github/actions/instrumentations/test
 
-  instrumentation-cypress-config:
-    runs-on: ubuntu-latest
-    permissions:
-      id-token: write
-    env:
-      PLUGINS: cypress-config
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
-      - uses: ./.github/actions/instrumentations/test
-
   instrumentation-express-mongo-sanitize:
     runs-on: ubuntu-latest
     permissions:
