+++ integration-tests/cypress/cypress-reporting-instrumentation.spec.js
@@ -6,8 +6,11 @@ const { once } = require('node:events')
 const fs = require('node:fs')
 const os = require('node:os')
 const path = require('node:path')
+const { format } = require('node:util')
 
+const proxyquire = require('proxyquire').noPreserveCache()
 const semver = require('semver')
+const sinon = require('sinon')
 const {
   sandboxCwd,
   useSandbox,
@@ -1469,3 +1472,407 @@ moduleTypes.forEach(({
     })
   })
 })
+
+// These filesystem failure tests do not depend on a Cypress version. Run them in one existing
+// integration matrix cell instead of repeating them for every supported version and module type.
+if (requestedVersion === 'latest' &&
+  (!process.env.CYPRESS_MODULE_TYPE || process.env.CYPRESS_MODULE_TYPE === 'commonJS')) {
+  describe('cypress config instrumentation', () => {
+    const temporaryDirectories = []
+    let errors
+
+    beforeEach(() => {
+      errors = []
+      sinon.stub(console, 'error').callsFake((...args) => errors.push(format(...args)))
+    })
+
+    afterEach(() => {
+      sinon.restore()
+      for (const directory of temporaryDirectories.splice(0)) {
+        fs.rmSync(directory, { force: true, recursive: true })
+      }
+    })
+
+    /**
+     * @param {string} code filesystem error code
+     * @param {string} filePath affected path
+     * @param {string} [syscall] failed system call
+     * @returns {NodeJS.ErrnoException} filesystem error
+     */
+    function createFileError (code, filePath, syscall = 'open') {
+      return Object.assign(new Error(`${code}: ${syscall} ${filePath}`), {
+        code,
+        path: filePath,
+        syscall,
+      })
+    }
+
+    /**
+     * @returns {string} temporary Cypress project root
+     */
+    function createProjectRoot () {
+      const projectRoot = fs.mkdtempSync(path.join(os.tmpdir(), 'dd-cypress-config-'))
+      temporaryDirectories.push(projectRoot)
+      return projectRoot
+    }
+
+    /**
+     * @param {object} [fsStub] filesystem overrides
+     * @param {() => string} [randomUUID] UUID generator
+     * @returns {{ cypressConfig: object, errors: string[], warnings: string[] }} loaded instrumentation and logs
+     */
+    function loadCypressConfig (fsStub, randomUUID) {
+      const warnings = []
+      let uuid = 0
+      const stubs = {
+        crypto: {
+          randomUUID: randomUUID || (() => `uuid-${++uuid}`),
+        },
+        '../../dd-trace/src/log': {
+          warn: (...args) => warnings.push(format(...args)),
+        },
+        './helpers/instrument': {
+          channel: () => ({ hasSubscribers: false, publish: () => {} }),
+        },
+      }
+
+      if (fsStub) stubs.fs = fsStub
+
+      return {
+        cypressConfig: proxyquire('../../packages/datadog-instrumentations/src/cypress-config', stubs),
+        errors,
+        warnings,
+      }
+    }
+
+    /**
+     * @param {object} cypressConfig Cypress config instrumentation
+     * @param {object} resolvedConfig resolved Cypress config
+     * @returns {{ handlers: Record<string, Function>, result: object }} registered handlers and returned config
+     */
+    function injectSupportFile (cypressConfig, resolvedConfig) {
+      const configFile = { e2e: {} }
+      const handlers = {}
+      cypressConfig.wrapConfig(configFile)
+
+      const result = configFile.e2e.setupNodeEvents((event, handler) => {
+        handlers[event] = handler
+      }, resolvedConfig)
+
+      return { handlers, result }
+    }
+
+    /**
+     * @param {string} directory directory to inspect
+     * @returns {string[]} generated Cypress files
+     */
+    function getGeneratedFiles (directory) {
+      return fs.readdirSync(directory)
+        .filter(file => file.startsWith('dd-cypress-support-') || file.startsWith('.dd-cypress-config-'))
+    }
+
+    /**
+     * @param {string} code error code raised after a partial write
+     * @param {(writeNumber: number) => boolean} [shouldFail] selects the write that fails
+     * @returns {object} filesystem stub
+     */
+    function createPartialWriteFailure (code, shouldFail = () => true) {
+      const pathsByDescriptor = new Map()
+      let writeNumber = 0
+
+      return {
+        openSync (filePath, flags) {
+          const descriptor = fs.openSync(filePath, flags)
+          pathsByDescriptor.set(descriptor, filePath)
+          return descriptor
+        },
+        writeFileSync (file, content, ...args) {
+          if (typeof file === 'number' && pathsByDescriptor.has(file)) {
+            writeNumber++
+            if (!shouldFail(writeNumber)) return fs.writeFileSync(file, content, ...args)
+
+            const filePath = pathsByDescriptor.get(file)
+            fs.writeFileSync(file, 'partial')
+            throw createFileError(code, filePath, 'write')
+          }
+          return fs.writeFileSync(file, content, ...args)
+        },
+        closeSync (descriptor) {
+          pathsByDescriptor.delete(descriptor)
+          return fs.closeSync(descriptor)
+        },
+      }
+    }
+
+    describe('support wrapper', () => {
+      it('falls back to the project root when the support directory is not writable', async () => {
+        const projectRoot = createProjectRoot()
+        const supportDirectory = path.join(projectRoot, 'cypress', 'support')
+        const supportFile = path.join(supportDirectory, 'e2e.js')
+        fs.mkdirSync(supportDirectory, { recursive: true })
+        fs.writeFileSync(supportFile, '// user support\n')
+
+        const { cypressConfig, warnings } = loadCypressConfig({
+          openSync (filePath, flags) {
+            if (path.dirname(filePath) === supportDirectory) {
+              throw createFileError('EACCES', filePath)
+            }
+            return fs.openSync(filePath, flags)
+          },
+        })
+        const resolvedConfig = { projectRoot, supportFile }
+        const { handlers } = injectSupportFile(cypressConfig, resolvedConfig)
+
+        assert.strictEqual(path.dirname(resolvedConfig.supportFile), projectRoot)
+        assert.deepStrictEqual(warnings, [])
+        assert.strictEqual(getGeneratedFiles(projectRoot).length, 2)
+
+        await handlers['after:run']({})
+        assert.deepStrictEqual(getGeneratedFiles(projectRoot), [])
+      })
+
+      it('logs an error with every failed location when no directory is writable', () => {
+        const projectRoot = createProjectRoot()
+        const supportDirectory = path.join(projectRoot, 'cypress', 'support')
+        const supportFile = path.join(supportDirectory, 'e2e.js')
+        fs.mkdirSync(supportDirectory, { recursive: true })
+        fs.writeFileSync(supportFile, '// user support\n')
+
+        const { cypressConfig, errors, warnings } = loadCypressConfig({
+          openSync (filePath) {
+            throw createFileError('EROFS', filePath)
+          },
+        })
+        const resolvedConfig = { projectRoot, supportFile }
+        injectSupportFile(cypressConfig, resolvedConfig)
+
+        assert.strictEqual(resolvedConfig.supportFile, supportFile)
+        assert.strictEqual(errors.length, 1)
+        assert.deepStrictEqual(warnings, [])
+        assert.match(errors[0], /^ERROR: Datadog could not create the Cypress support wrapper/)
+        assert.strictEqual(errors[0].match(/EROFS during open/g).length, 2)
+        assert.match(errors[0], new RegExp(supportDirectory.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')))
+        assert.match(errors[0], new RegExp(projectRoot.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')))
+      })
+
+      it('logs an error when the original support file cannot be read', () => {
+        const projectRoot = createProjectRoot()
+        const supportFile = path.join(projectRoot, 'e2e.js')
+        fs.writeFileSync(supportFile, '// user support\n')
+
+        const { cypressConfig, errors } = loadCypressConfig({
+          readFileSync (filePath, ...args) {
+            if (filePath === supportFile) throw createFileError('EACCES', filePath, 'read')
+            return fs.readFileSync(filePath, ...args)
+          },
+        })
+        const resolvedConfig = { projectRoot, supportFile }
+        injectSupportFile(cypressConfig, resolvedConfig)
+
+        assert.strictEqual(resolvedConfig.supportFile, supportFile)
+        assert.strictEqual(errors.length, 1)
+        assert.match(errors[0], /^ERROR: Datadog could not read the Cypress support file/)
+        assert.match(errors[0], /EACCES during read/)
+      })
+
+      it('logs an error when the Datadog browser hooks cannot be read', () => {
+        const projectRoot = createProjectRoot()
+        const browserHooksPath = require.resolve('../../packages/datadog-plugin-cypress/src/support')
+
+        const { cypressConfig, errors } = loadCypressConfig({
+          readFileSync (filePath, ...args) {
+            if (filePath === browserHooksPath) throw createFileError('EACCES', filePath, 'read')
+            return fs.readFileSync(filePath, ...args)
+          },
+        })
+        const resolvedConfig = { projectRoot, supportFile: false }
+        injectSupportFile(cypressConfig, resolvedConfig)
+
+        assert.strictEqual(resolvedConfig.supportFile, false)
+        assert.strictEqual(errors.length, 1)
+        assert.match(errors[0], /^ERROR: Datadog could not read its Cypress browser support hooks/)
+        assert.match(errors[0], /EACCES during read/)
+      })
+
+      it('logs an error when no project directory is available for the support wrapper', () => {
+        const { cypressConfig, errors } = loadCypressConfig()
+        const resolvedConfig = { supportFile: false }
+        injectSupportFile(cypressConfig, resolvedConfig)
+
+        assert.strictEqual(resolvedConfig.supportFile, false)
+        assert.strictEqual(errors.length, 1)
+        assert.match(errors[0], /^ERROR: Datadog could not create the Cypress support wrapper/)
+        assert.match(errors[0], /no project directory was available/)
+      })
+
+      it('removes partial support files when the filesystem runs out of space', () => {
+        const projectRoot = createProjectRoot()
+        const supportDirectory = path.join(projectRoot, 'cypress', 'support')
+        const supportFile = path.join(supportDirectory, 'e2e.js')
+        fs.mkdirSync(supportDirectory, { recursive: true })
+        fs.writeFileSync(supportFile, '// user support\n')
+
+        const { cypressConfig, errors } = loadCypressConfig(createPartialWriteFailure('ENOSPC'))
+        injectSupportFile(cypressConfig, { projectRoot, supportFile })
+
+        assert.strictEqual(errors.length, 1)
+        assert.match(errors[0], /ENOSPC during write/)
+        assert.deepStrictEqual(getGeneratedFiles(supportDirectory), [])
+        assert.deepStrictEqual(getGeneratedFiles(projectRoot), [])
+      })
+
+      it('removes the browser hooks when writing the support wrapper fails', () => {
+        const projectRoot = createProjectRoot()
+        const supportDirectory = path.join(projectRoot, 'cypress', 'support')
+        const supportFile = path.join(supportDirectory, 'e2e.js')
+        fs.mkdirSync(supportDirectory, { recursive: true })
+        fs.writeFileSync(supportFile, '// user support\n')
+
+        const failWrapperWrites = writeNumber => writeNumber % 2 === 0
+        const { cypressConfig, errors } = loadCypressConfig(
+          createPartialWriteFailure('ENOSPC', failWrapperWrites)
+        )
+        injectSupportFile(cypressConfig, { projectRoot, supportFile })
+
+        assert.strictEqual(errors.length, 1)
+        assert.strictEqual(errors[0].match(/ENOSPC during write/g).length, 2)
+        assert.deepStrictEqual(getGeneratedFiles(supportDirectory), [])
+        assert.deepStrictEqual(getGeneratedFiles(projectRoot), [])
+      })
+
+      it('removes partial support files when closing them fails', () => {
+        const projectRoot = createProjectRoot()
+        const supportDirectory = path.join(projectRoot, 'cypress', 'support')
+        const supportFile = path.join(supportDirectory, 'e2e.js')
+        const pathsByDescriptor = new Map()
+        fs.mkdirSync(supportDirectory, { recursive: true })
+        fs.writeFileSync(supportFile, '// user support\n')
+
+        const { cypressConfig, errors } = loadCypressConfig({
+          openSync (filePath, flags) {
+            const descriptor = fs.openSync(filePath, flags)
+            pathsByDescriptor.set(descriptor, filePath)
+            return descriptor
+          },
+          closeSync (descriptor) {
+            const filePath = pathsByDescriptor.get(descriptor)
+            pathsByDescriptor.delete(descriptor)
+            fs.closeSync(descriptor)
+            throw createFileError('EIO', filePath, 'close')
+          },
+        })
+        injectSupportFile(cypressConfig, { projectRoot, supportFile })
+
+        assert.strictEqual(errors.length, 1)
+        assert.strictEqual(errors[0].match(/EIO during close/g).length, 2)
+        assert.deepStrictEqual(getGeneratedFiles(supportDirectory), [])
+        assert.deepStrictEqual(getGeneratedFiles(projectRoot), [])
+      })
+
+      it('warns when generated support files cannot be removed', async () => {
+        const projectRoot = createProjectRoot()
+
+        const { cypressConfig, errors, warnings } = loadCypressConfig({
+          unlinkSync (filePath) {
+            throw createFileError('EACCES', filePath, 'unlink')
+          },
+        })
+        const { handlers } = injectSupportFile(cypressConfig, { projectRoot, supportFile: false })
+
+        await handlers['after:run']({})
+
+        assert.deepStrictEqual(errors, [])
+        assert.strictEqual(warnings.length, 2)
+        assert.ok(warnings.every(warning => warning.includes('could not remove generated Cypress file')))
+        assert.ok(warnings.every(warning => warning.includes('EACCES during unlink')))
+        assert.strictEqual(getGeneratedFiles(projectRoot).length, 2)
+      })
+    })
+
+    describe('configuration wrapper', () => {
+      it('falls back to the project root when the config directory is not writable', () => {
+        const projectRoot = createProjectRoot()
+        const configDirectory = path.join(projectRoot, 'config')
+        const configFile = path.join(configDirectory, 'cypress.config.js')
+        fs.mkdirSync(configDirectory)
+        fs.writeFileSync(configFile, 'module.exports = {}\n')
+
+        const { cypressConfig, warnings } = loadCypressConfig({
+          openSync (filePath, flags) {
+            if (path.dirname(filePath) === configDirectory) {
+              throw createFileError('EACCES', filePath)
+            }
+            return fs.openSync(filePath, flags)
+          },
+        })
+        const result = cypressConfig.wrapCliConfigFileOptions({
+          configFile,
+          project: projectRoot,
+        })
+
+        assert.strictEqual(path.dirname(result.options.configFile), projectRoot)
+        assert.deepStrictEqual(warnings, [])
+        assert.strictEqual(getGeneratedFiles(projectRoot).length, 1)
+
+        result.cleanup()
+        assert.deepStrictEqual(getGeneratedFiles(projectRoot), [])
+      })
+
+      it('warns with every failed location when no directory is writable', () => {
+        const projectRoot = createProjectRoot()
+        const configDirectory = path.join(projectRoot, 'config')
+        const configFile = path.join(configDirectory, 'cypress.config.js')
+        fs.mkdirSync(configDirectory)
+        fs.writeFileSync(configFile, 'module.exports = {}\n')
+
+        const { cypressConfig, errors, warnings } = loadCypressConfig({
+          openSync (filePath) {
+            throw createFileError('EROFS', filePath)
+          },
+        })
+        const options = { configFile, project: projectRoot }
+        const result = cypressConfig.wrapCliConfigFileOptions(options)
+
+        assert.strictEqual(result.options, options)
+        assert.deepStrictEqual(errors, [])
+        assert.strictEqual(warnings.length, 1)
+        assert.match(warnings[0], /could not create the Cypress configuration wrapper/)
+        assert.strictEqual(warnings[0].match(/EROFS during open/g).length, 2)
+        assert.match(warnings[0], new RegExp(configDirectory.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')))
+        assert.match(warnings[0], new RegExp(projectRoot.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')))
+      })
+
+      it('does not overwrite an existing configuration wrapper', () => {
+        const projectRoot = createProjectRoot()
+        const configFile = path.join(projectRoot, 'cypress.config.js')
+        const existingWrapper = path.join(projectRoot, `.dd-cypress-config-${process.pid}-collision.cjs`)
+        fs.writeFileSync(configFile, 'module.exports = {}\n')
+        fs.writeFileSync(existingWrapper, 'existing content\n')
+
+        const { cypressConfig, warnings } = loadCypressConfig(undefined, () => 'collision')
+        const options = { configFile, project: projectRoot }
+        const result = cypressConfig.wrapCliConfigFileOptions(options)
+
+        assert.strictEqual(result.options, options)
+        assert.strictEqual(fs.readFileSync(existingWrapper, 'utf8'), 'existing content\n')
+        assert.strictEqual(warnings.length, 1)
+        assert.match(warnings[0], /EEXIST during open/)
+      })
+
+      it('removes a partial configuration wrapper when the filesystem runs out of space', () => {
+        const projectRoot = createProjectRoot()
+        const configFile = path.join(projectRoot, 'cypress.config.js')
+        fs.writeFileSync(configFile, 'module.exports = {}\n')
+
+        const { cypressConfig, warnings } = loadCypressConfig(createPartialWriteFailure('ENOSPC'))
+        const options = { configFile, project: projectRoot }
+        const result = cypressConfig.wrapCliConfigFileOptions(options)
+
+        assert.strictEqual(result.options, options)
+        assert.strictEqual(warnings.length, 1)
+        assert.match(warnings[0], /ENOSPC during write/)
+        assert.deepStrictEqual(getGeneratedFiles(projectRoot), [])
+      })
+    })
+  })
+}
